requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
tensorflow==2.15.0
pillow>=9.5.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
